+++ package.json
@@ -44,6 +44,7 @@
     "react-dev-utils": "7.0.1",
     "react-dom": "16.8.1",
     "react-helmet": "5.2.0",
+    "react-router-dom": "5.0.0",
     "resolve": "1.8.1",
     "sass-loader": "7.1.0",
     "style-loader": "0.23.0",
