+++ src/Blog/BlogList.js
@@ -1,0 +1,5 @@
+import React, { Fragment } from "react";
+
+export default function BlogList() {
+  return <Fragment>Blog</Fragment>;
+}
